// fluentlyapp/lib/types/enums/Language.ts
export enum Language {
    Unknown = -1,
    English = 'English',
    Italian = 'Italian',
    German = 'German',
    French = 'French',
    Spanish = 'Spanish',
    Greek = 'Greek',
    Hindi = 'Hindi',
    Vietnamese = 'Vietnamese',
    Portugese = 'Portugese',
    Arabic = 'Arabic',
    Farsi = 'Farsi',
    Dutch = 'Dutch',
    Cantonese = 'Cantonese',
    ChineseSimplified = 'ChineseSimplified'
}